Introduce register_code/unregister_code
The purpose of these methods is to be able to register events without
having to figure out a physical key index for them. The press/release
methods are still there, and can be overridden for those cases where one
may want the index, but for everything else, there is register_code and
unregister_code, callable from user code.

+++ lib/Akela-TapDance/src/Akela/TapDance/Component/OneShotLayer.cpp
@@ -25,7 +25,7 @@ namespace Akela {
       void
       OneShotLayer::oneShotActivate (Akela::AbstractHID *,
                                      Akela::KeyMap *keymap,
-                                     uint8_t keycode) {
+                                     uint16_t keycode) {
         Akela::LayeredKeyMap *km = (Akela::LayeredKeyMap *)keymap;
         uint8_t layer = keycode - FN_ONESHOT_OSL;
         prev_layer = km->layer ();
@@ -36,7 +36,7 @@ namespace Akela {
       void
       OneShotLayer::oneShotDeactivate (Akela::AbstractHID *,
                                        Akela::KeyMap *keymap,
-                                       uint8_t) {
+                                       uint16_t) {
         Akela::LayeredKeyMap *km = (Akela::LayeredKeyMap *)keymap;
         km->layer (prev_layer);
         prev_layer = 0;
